Organize Static Features
This moves all custom Static Feature classes to a central folder. It also consolidates the Semantic Token manifest parsing helper methods and moves them to the CustomSemanticTokens class.
