+++ src/extension.ts
@@ -44,8 +44,6 @@ export async function activate(context: vscode.ExtensionContext): Promise<void> 
 
   const lsPath = new ServerPath(context);
   clientHandler = new ClientHandler(lsPath, outputChannel, reporter);
-  clientHandler.extSemanticTokenTypes = tokenTypesFromExtManifest(manifest);
-  clientHandler.extSemanticTokenModifiers = tokenModifiersFromExtManifest(manifest);
 
   // get rid of pre-2.0.0 settings
   if (config('terraform').has('languageServer.enabled')) {
@@ -312,32 +310,6 @@ async function terraformCommand(command: string, languageServerExec = true): Pro
     vscode.window.showWarningMessage(`Open a module then run terraform ${command} again`);
     return;
   }
-}
-
-interface PartialManifest {
-  contributes: {
-    semanticTokenTypes?: ObjectWithId[];
-    semanticTokenModifiers?: ObjectWithId[];
-  };
-}
-
-interface ObjectWithId {
-  id: string;
-}
-
-function tokenTypesFromExtManifest(manifest: PartialManifest): string[] {
-  if (!manifest.contributes.semanticTokenTypes) {
-    return [];
-  }
-  return manifest.contributes.semanticTokenTypes.map((token: ObjectWithId) => token.id);
-}
-
-function tokenModifiersFromExtManifest(manifest: PartialManifest): string[] {
-  if (!manifest.contributes.semanticTokenModifiers) {
-    return [];
-  }
-
-  return manifest.contributes.semanticTokenModifiers.map((modifier: ObjectWithId) => modifier.id);
 }
 
 function enabled(): boolean {
