+++ src/features/semanticTokens.ts
@@ -1,0 +1,55 @@
+import { BaseLanguageClient, ClientCapabilities, ServerCapabilities, StaticFeature } from 'vscode-languageclient';
+
+interface PartialManifest {
+  contributes: {
+    semanticTokenTypes?: ObjectWithId[];
+    semanticTokenModifiers?: ObjectWithId[];
+  };
+}
+
+interface ObjectWithId {
+  id: string;
+}
+
+export class CustomSemanticTokens implements StaticFeature {
+  constructor(private _client: BaseLanguageClient, private manifest: PartialManifest) {}
+
+  public fillClientCapabilities(capabilities: ClientCapabilities): void {
+    if (!capabilities.textDocument || !capabilities.textDocument.semanticTokens) {
+      return;
+    }
+
+    const extSemanticTokenTypes = this.tokenTypesFromExtManifest(this.manifest);
+    const extSemanticTokenModifiers = this.tokenModifiersFromExtManifest(this.manifest);
+
+    const tokenTypes = capabilities.textDocument.semanticTokens.tokenTypes;
+    capabilities.textDocument.semanticTokens.tokenTypes = tokenTypes.concat(extSemanticTokenTypes);
+
+    const tokenModifiers = capabilities.textDocument.semanticTokens.tokenModifiers;
+    capabilities.textDocument.semanticTokens.tokenModifiers = tokenModifiers.concat(extSemanticTokenModifiers);
+  }
+
+  // eslint-disable-next-line @typescript-eslint/no-unused-vars
+  public initialize(capabilities: ServerCapabilities): void {
+    return;
+  }
+
+  public dispose(): void {
+    return;
+  }
+
+  tokenTypesFromExtManifest(manifest: PartialManifest): string[] {
+    if (!manifest.contributes.semanticTokenTypes) {
+      return [];
+    }
+    return manifest.contributes.semanticTokenTypes.map((token: ObjectWithId) => token.id);
+  }
+
+  tokenModifiersFromExtManifest(manifest: PartialManifest): string[] {
+    if (!manifest.contributes.semanticTokenModifiers) {
+      return [];
+    }
+
+    return manifest.contributes.semanticTokenModifiers.map((modifier: ObjectWithId) => modifier.id);
+  }
+}
